requirements.txt
advanced-alchemy==0.14.1
alembic==1.13.1
annotated-types==0.7.0
anyio==4.3.0
cachetools==5.3.3
certifi==2024.2.2
click==8.1.7
editorconfig==0.12.4
faker==25.2.0
google-api-core==2.19.0
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
jinja2==3.1.4
jsbeautifier==1.15.1
litestar==2.8.3
mako==1.3.5
markdown-it-py==3.0.0
mdurl==0.1.2
oauthlib==3.2.2
polyfactory==2.16.0
proto-plus==1.24.0
pyasn1==0.6.0
pyasn1-modules==0.4.0
pydantic==2.7.4
pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rich-click==1.8.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
typing-extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.29.0
